Fix warnings on windows 64bit

+++ libraries/entities/src/EntityTree.cpp
@@ -1055,7 +1055,7 @@ int EntityTree::processEraseMessageDetails(const QByteArray& dataByteArray, cons
                 break; // bail to prevent buffer overflow
             }
 
-            QByteArray encodedID = dataByteArray.mid(processedBytes, NUM_BYTES_RFC4122_UUID);
+            QByteArray encodedID = dataByteArray.mid((int)processedBytes, NUM_BYTES_RFC4122_UUID);
             QUuid entityID = QUuid::fromRfc4122(encodedID);
             dataAt += encodedID.size();
             processedBytes += encodedID.size();
@@ -1074,7 +1074,7 @@ int EntityTree::processEraseMessageDetails(const QByteArray& dataByteArray, cons
         }
         deleteEntities(entityItemIDsToDelete, true, true);
     }
-    return processedBytes;
+    return (int)processedBytes;
 }
 
 EntityTreeElementPointer EntityTree::getContainingElement(const EntityItemID& entityItemID)  /*const*/ {
